[BEAM-944] Spark runner causes an exception when creating pipeline options. Create a SparkContextOptions for context-ware options.
