+++ runners/spark/src/main/java/org/apache/beam/runners/spark/SparkPipelineOptions.java
@@ -18,24 +18,22 @@
 
 package org.apache.beam.runners.spark;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import java.util.ArrayList;
-import java.util.List;
 import org.apache.beam.sdk.options.ApplicationNameOptions;
 import org.apache.beam.sdk.options.Default;
 import org.apache.beam.sdk.options.DefaultValueFactory;
 import org.apache.beam.sdk.options.Description;
 import org.apache.beam.sdk.options.PipelineOptions;
 import org.apache.beam.sdk.options.StreamingOptions;
-import org.apache.spark.api.java.JavaSparkContext;
-import org.apache.spark.streaming.api.java.JavaStreamingListener;
+
 
 
 /**
- * Spark runner pipeline options.
+ * Spark runner {@link PipelineOptions} handles Spark execution-related configurations,
+ * such as the master address, batch-interval, and other user-related knobs.
  */
-public interface SparkPipelineOptions extends PipelineOptions, StreamingOptions,
-                                              ApplicationNameOptions {
+public interface SparkPipelineOptions
+    extends PipelineOptions, StreamingOptions, ApplicationNameOptions {
+
   @Description("The url of the spark master to connect to, (e.g. spark://host:port, local[4]).")
   @Default.String("local[4]")
   String getSparkMaster();
@@ -92,28 +90,4 @@ public interface SparkPipelineOptions extends PipelineOptions, StreamingOptions,
   @Default.Boolean(true)
   Boolean getEnableSparkSinks();
   void setEnableSparkSinks(Boolean enableSparkSinks);
-
-  @Description("If the spark runner will be initialized with a provided Spark Context")
-  @Default.Boolean(false)
-  boolean getUsesProvidedSparkContext();
-  void setUsesProvidedSparkContext(boolean value);
-
-  @Description("Provided Java Spark Context")
-  @JsonIgnore
-  JavaSparkContext getProvidedSparkContext();
-  void setProvidedSparkContext(JavaSparkContext jsc);
-
-  @Description("Spark streaming listeners")
-  @Default.InstanceFactory(EmptyListenersList.class)
-  @JsonIgnore
-  List<JavaStreamingListener> getListeners();
-  void setListeners(List<JavaStreamingListener> listeners);
-
-  /** Returns an empty list, top avoid handling null. */
-  class EmptyListenersList implements DefaultValueFactory<List<JavaStreamingListener>> {
-    @Override
-    public List<JavaStreamingListener> create(PipelineOptions options) {
-      return new ArrayList<>();
-    }
-  }
 }
