+++ runners/spark/src/main/java/org/apache/beam/runners/spark/translation/SparkContextFactory.java
@@ -18,6 +18,7 @@
 
 package org.apache.beam.runners.spark.translation;
 
+import org.apache.beam.runners.spark.SparkContextOptions;
 import org.apache.beam.runners.spark.SparkPipelineOptions;
 import org.apache.beam.runners.spark.coders.BeamSparkRunnerRegistrator;
 import org.apache.spark.SparkConf;
@@ -46,11 +47,13 @@ public final class SparkContextFactory {
   }
 
   public static synchronized JavaSparkContext getSparkContext(SparkPipelineOptions options) {
+    SparkContextOptions contextOptions = options.as(SparkContextOptions.class);
     // reuse should be ignored if the context is provided.
-    if (Boolean.getBoolean(TEST_REUSE_SPARK_CONTEXT) && !options.getUsesProvidedSparkContext()) {
+    if (Boolean.getBoolean(TEST_REUSE_SPARK_CONTEXT)
+        && !contextOptions.getUsesProvidedSparkContext()) {
       // if the context is null or stopped for some reason, re-create it.
       if (sparkContext == null || sparkContext.sc().isStopped()) {
-        sparkContext = createSparkContext(options);
+        sparkContext = createSparkContext(contextOptions);
         sparkMaster = options.getSparkMaster();
       } else if (!options.getSparkMaster().equals(sparkMaster)) {
         throw new IllegalArgumentException(String.format("Cannot reuse spark context "
@@ -59,7 +62,7 @@ public final class SparkContextFactory {
       }
       return sparkContext;
     } else {
-      return createSparkContext(options);
+      return createSparkContext(contextOptions);
     }
   }
 
@@ -69,10 +72,10 @@ public final class SparkContextFactory {
     }
   }
 
-  private static JavaSparkContext createSparkContext(SparkPipelineOptions options) {
-    if (options.getUsesProvidedSparkContext()) {
+  private static JavaSparkContext createSparkContext(SparkContextOptions contextOptions) {
+    if (contextOptions.getUsesProvidedSparkContext()) {
       LOG.info("Using a provided Spark Context");
-      JavaSparkContext jsc = options.getProvidedSparkContext();
+      JavaSparkContext jsc = contextOptions.getProvidedSparkContext();
       if (jsc == null || jsc.sc().isStopped()){
         LOG.error("The provided Spark context " + jsc + " was not created or was stopped");
         throw new RuntimeException("The provided Spark context was not created or was stopped");
@@ -83,9 +86,9 @@ public final class SparkContextFactory {
       SparkConf conf = new SparkConf();
       if (!conf.contains("spark.master")) {
         // set master if not set.
-        conf.setMaster(options.getSparkMaster());
+        conf.setMaster(contextOptions.getSparkMaster());
       }
-      conf.setAppName(options.getAppName());
+      conf.setAppName(contextOptions.getAppName());
       // register immutable collections serializers because the SDK uses them.
       conf.set("spark.kryo.registrator", BeamSparkRunnerRegistrator.class.getName());
       conf.set("spark.serializer", KryoSerializer.class.getName());
